requirements.txt
requests==2.26.0
MarkupSafe==2.0.1
itsdangerous==2.0.1
Werkzeug==2.0.2
wheel==0.38.2
pandas==1.3.4
Flask==1.1.2
Jinja2==2.11.3
mysql-connector-python==8.0.31